[Feat] css ver1.2

+++ src/main/java/eat_schedule/dto/Store.java
@@ -1,7 +1,10 @@
 package eat_schedule.dto;
 
+import eat_schedule.mapper.CommonMapper;
 import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import lombok.Setter;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 @Component
@@ -12,7 +15,7 @@ public class Store {
 
     private String category;
 
-    private Integer score;
+    private Double score;
 
     private Integer review;
 
@@ -43,6 +46,7 @@ public class Store {
     private boolean playroom;
 
     private boolean disabled;
+
 
     public void setLocation(String location) {
         this.location = "\'"+location+"\'";
